Using lxml and Python for web scraping
This is the first post in a two post series on web scraping using lxml in Python. In this post I created the lxml based script. In the next post I will discuss how to convert this script into a Flask based API and how to host it on Heroku.